Added comments in main class you should read.
We should use inorder() and backward() methods in order to display
recommendations rather than while loops. 

I feel as if the mistake we are making lies within the getLeft() method.
That is, we are not envisioning the tree correctly, and left of 64 for
example, is actually nothing. therefore it does not work like we expect.


inorder(BSTNode r) will take found_Node and display all nodes (inorder)
from found_Node to the rightmode node in tree (highest value)

backwards(BSTNode r) will use inorder() algorithm, however it will
take the output, and reverse it.